Some graphics updates
ADD: Some buttons for each room.
ADD: Each room background is now an animation, it just need to be drawn
!
Update: The room layout slightly changed in order to incrase the size of
the smallest rooms.
ADD: Message can be easily added to the bottom part of each room.
ADD: Status bar for the repairable rooms.