V0.5.0 Fully operationnal fights
ADD: Sounds for battles.
ADD: Ships with fights info for the enemy and rooms for the player.
ADD: Laser and rockets animations, actions and effects.
Update: Repairable room now have integers status and the repairCost is
for 1% of the status.
Update: Buttons are in a table not in an HorizontalGroup anymore. Its
easier to config with the loading effect.
Update: Star class is not in PilotRoom anymore.