V0.6.1 Fixes
Fix: Sound was not desactivated from OptionScreen.
Update: Extract a method to fin the closest star from MoveToClosestStar.
It's used to check if there is enough resources to travel.
Update: Travel's cost is calculated from the ControlRoom and not
directly in travel method.
Update: Add a '+' in front of the collected amount when
harvesting.resources.